requirements.txt
alabaster==0.7.16
appdirs==1.4.4
babel==2.16.0
boto3==1.35.83
botocore==1.35.83
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
Deprecated==1.2.15
docutils==0.21.2
fonttools==4.55.3
gevent==24.11.1
greenlet==3.1.1
idna==3.10
imagesize==1.4.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.4
jmespath==1.0.1
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdit-py-plugins==0.4.2
mdurl==0.1.2
mpl-finance==0.10.1
mplcursors==0.6
myst-parser==3.0.1
numpy==2.0.2
packaging==24.2
pandas==2.2.3
pdoc==15.0.1
pillow==11.0.0
pycparser==2.22
PyGithub==2.5.0
Pygments==2.18.0
PyJWT==2.10.1
PyNaCl==1.5.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-lambda-local==0.1.13
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rinoh-typeface-dejavuserif==0.1.3
rinoh-typeface-texgyrecursor==0.1.1
rinoh-typeface-texgyreheros==0.1.1
rinoh-typeface-texgyrepagella==0.1.1
rinohtype==0.5.5
s3transfer==0.10.4
six==1.17.0
snowballstemmer==2.2.0
Sphinx==7.4.7
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sphinxemoji==0.3.1
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
websocket-client==1.8.0
wrapt==1.17.0
zipp==3.21.0
zope.event==5.0
zope.interface==7.2
